UC-12 refactor: optimize CircleBlocks.tsx
