+++ src/entities/Posts/lib/index.ts
@@ -1,2 +1,1 @@
 export * from './truncateDescription'
-export * from './useCircleBlocks'
